+++ MegaManX3/AnimationOneTime.cpp
@@ -31,21 +31,21 @@ AnimationOneTime::AnimationOneTime(int defaultTime) : CAnimation(defaultTime)
 	isDone = false;
 }
 
-void AnimationOneTime::render(int x, int y, bool center, D3DCOLOR colorBrush)
+void AnimationOneTime::render(float x, float y, bool center, D3DCOLOR colorBrush)
 {
 	if (!isDone)
 		_updateFrame(),
 		frames[currentFrame]->getSprite()->draw(x, y, center, colorBrush);
 }
 
-void AnimationOneTime::renderFlipX(int x, int y, bool center, D3DCOLOR colorBrush)
+void AnimationOneTime::renderFlipX(float x, float y, bool center, D3DCOLOR colorBrush)
 {
 	if (!isDone)
 		_updateFrame(),
 		frames[currentFrame]->getSprite()->drawFlipX(x, y, center, colorBrush);
 }
 
-void AnimationOneTime::renderFlipY(int x, int y, bool center, D3DCOLOR colorBrush)
+void AnimationOneTime::renderFlipY(float x, float y, bool center, D3DCOLOR colorBrush)
 {
 	if (!isDone)
 		_updateFrame(),
